python_ecg_lava_api/requirements.txt
fastapi>=0.111.0
httpx[http2]>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
uvicorn>=0.30.0
